% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BarsAndSets.R
\name{BarsAndSetsPlot}
\alias{BarsAndSetsPlot}
\title{Plot omics data along with functional classification}
\usage{
BarsAndSetsPlot(bar_values, categories)
}
\arguments{
\item{bar_values}{log transformed, ordered vector of ratios.}

\item{categories}{Dataframe with T/F values indicating the membership in categories
and column names as labels.}
}
\description{
This function plots a dataset (e.g. gene expression ratios) as a bar plot along with
one or multiple tracks indicating the membership in functional or regulatory categories.
}
\details{
Expression data is given in the first argument as a vector of ratios. Note that the data
should be log transformed and ordered ascending for best results.

Functional categories are given as a dataframe a column of TRUE/FALSE values for each set.
Column names of the data frame are used as labels in the plot.
}
